requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
django-environ==0.11.2

# Authentication
django-allauth==0.57.0

# Cache & Session
redis==5.0.1
django-redis==5.4.0

# AI & LangChain
langchain==0.0.350
langchain-openai==0.0.2
openai>=1.6.1,<2.0.0

# Background Tasks
celery==5.3.4
django-celery-beat==2.5.0

# API Documentation
drf-yasg==1.21.7

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-decouple==3.8
Pillow==10.1.0
django-filter==23.5
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# Monitoring & Logging
sentry-sdk==1.38.0
django-debug-toolbar==4.2.0
drf-yasg

# Security
django-ratelimit==4.1.0 